Atomic Design añadido para contactos
Se agrego la vista de contacto utilizando el patrón Atomic Design y se hicieron algunos cambios en el diseño de la app
