+++ src/features/contacs/presentation/ui/components/templates/ContacsTemplate.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import SafeLayout from '@/shared/components/layouts/SafeLayout';
+
+export default function ContacsTemplate({ children }: { children: React.ReactNode }) {
+    return (
+        <SafeLayout>
+            {children}
+        </SafeLayout>
+    );
+}
