+++ src/features/home/presentation/ui/components/organisms/RecentNotesCard.tsx
@@ -1,25 +1,50 @@
 import React from "react";
-import { StyleSheet, Text, View, Image } from "react-native";
+import { StyleSheet, Text, View, Image, Pressable } from "react-native";
 import { colors, spacing } from '@/shared/theme'; //Hoja de Estilos generales
 
 export default function RecentNotesCard () {
     return (
-        <View style={styles.noteCard}>
-            <View style={styles.cardBody}>
-                <View style={styles.imgContainer}>
-                    <Image 
-                        source={require('../../../../../../../assets/photo_prototype.png')}
-                        style={styles.contactImg}
-                    />
-                </View>
-                <View style={styles.contactInfo}>
-                    <View style={styles.infoHeader}>
-                        <Text>*Nombre</Text>
-                        <Text>*Fecha - Creación </Text>
-                    </View>
-                    <Text>*Nota</Text>
-                </View>
-            </View> 
+        <View>
+            <View style={styles.noteCard}>
+                <Pressable onPress={()=> console.log("Clic en la nota del contacto")}>
+                    {({pressed}) => (
+                        <View style={[styles.cardBody, pressed && styles.pressedCardBody]}>
+                            <View style={styles.imgContainer}>
+                                <Image 
+                                    source={require('../../../../../../../assets/photo_prototype.png')}
+                                    style={styles.contactImg}
+                                />
+                            </View>
+                            <View style={styles.contactInfo}>
+                                <View style={styles.infoHeader}>
+                                    <Text>*Nombre</Text>
+                                    <Text>*Fecha - Creación </Text>
+                                </View>
+                                <Text>*Nota</Text>
+                            </View>
+                        </View>
+                    )}
+                </Pressable>
+                <Pressable onPress={()=> console.log("Clic en la nota del contacto")}>
+                    {({pressed}) => (
+                        <View style={[styles.cardBody, pressed && styles.pressedCardBody]}>
+                            <View style={styles.imgContainer}>
+                                <Image 
+                                    source={require('../../../../../../../assets/photo_prototype.png')}
+                                    style={styles.contactImg}
+                                />
+                            </View>
+                            <View style={styles.contactInfo}>
+                                <View style={styles.infoHeader}>
+                                    <Text>*Nombre</Text>
+                                    <Text>*Fecha - Creación </Text>
+                                </View>
+                                <Text>*Nota</Text>
+                            </View>
+                        </View>
+                    )}
+                </Pressable>
+            </View>
         </View>
     );
 }
@@ -29,12 +54,17 @@ const styles = StyleSheet.create({
     noteCard:{
         backgroundColor: colors.surface,
         borderRadius:spacing.md,
-        paddingTop: spacing.md,
-        paddingBottom: spacing.md,
-        paddingLeft: 20,
+        paddingTop: spacing.sm,
+        paddingBottom: spacing.sm,
+        paddingLeft: spacing.none,
     },
     //Cuerpo de la tarjeta
-    cardBody:{flexDirection:'row'},
+    cardBody:{
+        flexDirection:'row', 
+        paddingLeft:spacing.md,
+        paddingBottom: spacing.sm,
+    },
+    pressedCardBody:{backgroundColor: colors.btnPressed},
     imgContainer:{
         justifyContent:'center',
     },
