+++ src/shared/navigation/AppNavigator.tsx
@@ -2,8 +2,8 @@
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { getTabOptions } from './TabOptions'; //función modularizada para la configuración del Tabs.Navigator
 import homePage from '../../features/home/presentation/ui/components/pages/HomePage';
-import dashPage from '../../features/dashboard/presentation/ui/page/DashPage';
-import contacsPage from '../../features/contacs/presentation/ui/page/ContacsPage';
+import dashPage from '../../features/dashboard/presentation/ui/components/pages/DashPage';
+import contacsPage from '../../features/contacs/presentation/ui/components/pages/ContacsPage';
 
 const Tabs = createBottomTabNavigator();
