+++ src/shared/theme/Colors.ts
@@ -1,9 +1,11 @@
 export const colors = {
     // colores más utilizados en toda la app
-    background: '#eeeeee',     // fondo general de pantallas
-    backgroundApp: 'tomato',   // fondo de toda la app
-    surface: 'white',          // botones, tarjetas
+    background: '#eeeeee', // fondo general de pantallas
+    backgroundApp: 'tomato', // fondo de toda la app
+    surface: 'white', // botones, tarjetas
+    btnPressed: "#e3e3e3", // Efecto hover en botones(Pressable)
+    
     // colores de Texto con frecuencia 
-    titleText: '#0089e8',      // títulos
-    defaultColor: 'black',      // texto por defecto
+    titleText: '#0089e8', // títulos
+    defaultColor: 'black', // texto por defecto
 } as const
